Transferred components from filterCards.jsx to just home.jsx instead of separating pages. #2
